% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%+=\%}
\alias{\%+=\%}
\title{Incremental function}
\usage{
t \%+=\% s
}
\arguments{
\item{t}{summand 1}

\item{s}{summand 2}
}
\description{
This function increments a variable by a given value. That's it.
}
\examples{
x <- 1
x \%+=\% 3
x
#> [1] 4
}
