% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oma.R
\name{oma}
\alias{oma}
\title{Optimal Mate Allocation}
\usage{
oma(data, K, F.max, solver = "ECOS")
}
\arguments{
\item{data}{Input data frame (see Details)}

\item{K}{kinship matrix to control inbreeding}

\item{F.max}{maximum inbreeding coefficient (vector allowed)}

\item{solver}{solver for CVXR (default is "ECOS")}
}
\value{
list containing
\describe{
\item{response}{data frame with F and merit}
\item{oc}{matrix of optimal contributions for each individual}
\item{om}{matrix of optimal allocations for each mating}
}
}
\description{
Optimize the allocation for each mating
}
\details{
The data.frame \code{data} has five columns: mother, father, min, max, merit
}
