% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ribbon.R
\name{plot_ribbon}
\alias{plot_ribbon}
\title{Ribbon Plot of Optimal Contributions}
\usage{
plot_ribbon(oc, min.c = 0.001)
}
\arguments{
\item{oc}{named list of optimal contributions}

\item{min.c}{minimum contribution for plotting}
}
\value{
ggplot2 object
}
\description{
Ribbon Plot of Optimal Contributions
}
\details{
The inbreeding rates are taken from the names of \code{oc}. Each element of \code{oc} should be a data frame with columns "id" and "value".
}
