% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data for OCS/OMA}
\usage{
read_data(
  geno.file,
  kinship.file,
  ploidy,
  sex = NULL,
  matings = "none",
  standardize = FALSE,
  n.core = 1
)
}
\arguments{
\item{geno.file}{file with marker effects and genotypes}

\item{kinship.file}{file with kinship matrix}

\item{ploidy}{even integer}

\item{sex}{optional, data frame with columns id and female (T/F)}

\item{matings}{see Details}

\item{standardize}{T/F, standardize merit in parental candidates}

\item{n.core}{multi-core evaluation}
}
\value{
list containing
\describe{
\item{K}{kinship matrix}
\item{parents}{data frame of individual merits}
\item{matings}{data frame of mating merits}
}
}
\description{
Predicts merit for OCS/OMA
}
\details{
The first column of \code{geno.file} is the marker name. The second column contains the additive effects for the breeding value parameterization, and (digenic) dominance effects (when available) should be in the third column with the header "dom". Subsequent columns contain the marker data for the population, coded as allele dosage, from 0 to ploidy. Missing marker data is imputed with the population mean.

The \code{kinship.file} should contain an N x N kinship matrix with id names in the first column and row.

There are several options for argument \code{matings}: (1) "none" = no matings; (2) "all" = all possible matings of the individuals in \code{geno.file} (excluding reciprocals); (3) a character vector of genotype ids to calculate all pairs of matings; (4) a 2-column data.frame of desired matings with header "female","male" for dioecious species (separate sexes) or "parent1","parent2" for hermaphrodites. Self-matings are included under options (2) and (3) for dioecious species but are easily removed in the output if desired.
}
