% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs.R
\name{ocs}
\alias{ocs}
\title{Optimal Contribution Selection}
\usage{
ocs(data, K, F.max, solver = "ECOS")
}
\arguments{
\item{data}{Input data frame (see Details)}

\item{K}{kinship matrix to control inbreeding}

\item{F.max}{maximum inbreeding coefficient (vector allowed)}

\item{solver}{solver for CVXR (default is "ECOS")}
}
\value{
list containing
\describe{
\item{response}{data frame with F and merit}
\item{oc}{matrix of optimal contributions for each F.max}
}
}
\description{
Optimize the contribution of each individual to the next generation
}
\details{
The data.frame \code{data} has four mandatory columns: id, merit, min, max. An optional fifth column specifies the sex of the individual as "M" (male) or "F" (female).
}
