% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs.R
\name{ocs}
\alias{ocs}
\title{Optimal Contribution Selection}
\usage{
ocs(parents, ploidy, K, dF, solver = "ECOS")
}
\arguments{
\item{parents}{input data frame (see Details)}

\item{ploidy}{ploidy}

\item{K}{kinship matrix}

\item{dF}{inbreeding rate (can be vector of values)}

\item{solver}{solver for CVXR (default is "ECOS")}
}
\value{
list containing
\describe{
\item{response}{data frame with F and merit}
\item{oc}{matrix of optimal contributions for each F.max}
}
}
\description{
Optimize the contribution of each individual to the next generation
}
\details{
The data.frame \code{parents} can have up to five columns: id, merit, min, max, female. Min and max are real numbers between 0 and 1 specifying the minimum and maximum contribution for each parent. The "female" column is logical TRUE/FALSE. For hermaphroditic species, the last column is omitted.

The average inbreeding coefficient of the current generation is based on all individuals in \code{K}.
}
