% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs.R
\name{ocs}
\alias{ocs}
\title{Optimal Contribution Selection}
\usage{
ocs(dF, parents, ploidy, K, tol = 1e-06, dF.adapt = NULL, solver = "ECOS")
}
\arguments{
\item{dF}{inbreeding rate}

\item{parents}{input data frame (see Details)}

\item{ploidy}{ploidy}

\item{K}{kinship matrix}

\item{tol}{tolerance, values below this set to 0}

\item{dF.adapt}{see Details}

\item{solver}{solver for CVXR (default is "ECOS")}
}
\value{
list containing
\describe{
\item{response}{data.frame with realized dF, merit, n.parent}
\item{oc}{data frame of optimal contributions}
}
}
\description{
Optimize the contribution of each individual to the next generation
}
\details{
The first four columns of \code{parents} should be named as follows: id, merit, min, max. Min and max are real numbers between 0 and 1 specifying the minimum and maximum contribution for each parent. An optional fifth column named "female" is a logical TRUE/FALSE variable for species with separate sexes. 

Additional columns can be used to specify constraints on linear combinations of the contributions. The values are the coefficients of the contribution variables, and the name of each column specifies the right-hand side of the constraint. Each name must begin with "lt","gt", or "eq", followed by a non-negative numeric value. For example, "lt0.5" means less than or equal to 0.5.

The average inbreeding coefficient of the current generation is based on all individuals in \code{K}, which may exceed the list of individuals in \code{parents}.

It is possible that no feasible solution exists for the specified \code{dF}. Argument \code{dF.adapt} can be used to automatically increase dF by dF.adapt$step up to dF.adapt$max.
}
