% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oma.R
\name{oma}
\alias{oma}
\title{Optimal Mate Allocation}
\usage{
oma(parents, matings, ploidy, K, dF, min.c = 0, solver = "ECOS")
}
\arguments{
\item{parents}{parents data frame (see Details)}

\item{matings}{matings data frame (see Details)}

\item{ploidy}{ploidy}

\item{K}{kinship matrix}

\item{dF}{inbreeding rate (can be vector of values)}

\item{min.c}{minimum allocation}

\item{solver}{solver for CVXR (default is "ECOS")}
}
\value{
list containing
\describe{
\item{response}{data frame with F and merit}
\item{oc}{matrix of optimal contributions for each individual}
\item{om}{matrix of optimal allocations for each mating}
}
}
\description{
Optimize the allocation for each mating
}
\details{
The data frame \code{parents} needs columns id, min, max, with optional column female (logical) for separate sexes. The data.frame \code{matings} has five columns: female, male, merit, min, max. 

After optimization, allocations less than \code{min.c} are set equal to zero, and the remainder are renormalized.
}
