% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_opt.R
\name{plot_opt}
\alias{plot_opt}
\title{Plot Optimal Contributions}
\usage{
plot_opt(data, min.c)
}
\arguments{
\item{data}{matrix of optimal contributions}

\item{min.c}{minimum contribution for plotting}
}
\value{
ggplot2 object
}
\description{
Plot Optimal Contributions
}
