% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data for OCS/OMA}
\usage{
read_data(geno.file, ploidy, dominance, sex = NULL, matings = NULL, n.core = 1)
}
\arguments{
\item{geno.file}{file with marker effects and genotypes}

\item{ploidy}{even integer}

\item{dominance}{TRUE/FALSE}

\item{sex}{optional, data frame with columns id and sex ("M" or "F")}

\item{matings}{data frame with four columns: mother, father, lower, upper}

\item{n.core}{multi-core evaluation}
}
\value{
list containing
\describe{
\item{K}{genomic kinship matrix}
\item{ocs.data}{data frame with individual merits and limits}
\item{oma.data}{data frame with mating merits and limits}
}
}
\description{
Constructs the G matrix and predicts merit for OCS/OMA
}
\details{
The first column of \code{geno.file} is marker name. The second column contains the additive effects for the breeding value parameterization. When \code{dominance=TRUE}, the third column contains (digenic) dominance effects. Subsequent columns contain the marker data for the population, coded as allele dosage, from 0 to ploidy. 

When \code{matings} is present, merit is predicted for all matings in the data frame. The columns "lower" and "upper" specify the limits (0-1) for the contribution of each mating.
}
