% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data for OCS/OMA}
\usage{
read_data(
  geno.file,
  kinship.file = NULL,
  ploidy,
  dominance,
  sex = NULL,
  matings,
  n.core = 1
)
}
\arguments{
\item{geno.file}{file with marker effects and genotypes}

\item{kinship.file}{NULL}

\item{ploidy}{even integer}

\item{dominance}{TRUE/FALSE}

\item{sex}{optional, data frame with columns id (character) and female (TRUE/FALSE)}

\item{matings}{see Details}

\item{n.core}{multi-core evaluation}
}
\value{
list containing
\describe{
\item{K}{genomic kinship matrix}
\item{parents}{data frame with individual merits and limits}
\item{matings}{data frame with mating merits and limits}
}
}
\description{
Constructs the G matrix and predicts merit for OCS/OMA
}
\details{
The first column of \code{geno.file} is marker name. The second column contains the additive effects for the breeding value parameterization, and when \code{dominance=TRUE}, the next column contains (digenic) dominance effects. The following column "p.ref" should contain allele frequencies for the reference population, which is needed to control genomic inbreeding. Subsequent columns contain the marker data for the population, coded as allele dosage, from 0 to ploidy. 

The argument \code{matings} can be either (1) a character vector of genotype ids, in which case all possible matings of these individuals are generated (reciprocal matings excluded), or (2) a data.frame with four columns: female, male, min, max. The columns "min" and "max" specify the limits (0-1) for the contribution of each mating. To skip mate allocation entirely, use \code{matings=NULL}.
}
